Fix close button on media view for iPhone X
Currently, after clicking a photo or video from a timeline tweet,
you can not close the media view controller on the iPhone X. The
close button overlaps the status bar and does not respond to user
interaction. This code uses the safe area layout guide in iOS 11 to
constrain the top of the close button to the top of the safe area.